feat: Update visualization functions
- move plotting functions into src/utils/plot_utils.py
- add plot function for testing
